Feat: Move the Models into their own file.
Also seperate the app inititlization from the handlers.
