+++ api/__init__.py
@@ -1,7 +1,12 @@
+import os
+
 import connexion
-import requests
 
-connexion_app = connexion.App(__name__)
+app = connexion.App(__name__, specification_dir='./')
 
-from item_tracker_core import handlers
-
+app.add_api('openapi.yml')
+ENV = os.getenv('FYS_WORKING_ENV')
+if ENV:
+    if ENV.lower() == 'dev':
+        os.environ['FLASK_ENV'] = 'development'
+        app.run(port=8080, debug=True)
