+++ api/handlers.py
@@ -1,14 +1,11 @@
 import os
-import psycopg2
-import uuid
+
 import connexion
-from sqlalchemy import create_engine, Table, Column, String, MetaData, Integer, Boolean, Sequence, BigInteger, ForeignKey, DateTime
-from sqlalchemy.ext.declarative import declarative_base
+
+from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from sqlalchemy.dialects.postgresql import UUID
 
-
-#$Env:DATABASE_URL = $(heroku config:get DATABASE_URL -a found-your-stuff-api);  py.exe handlers.py
+from models import User,Tag
 
 if not os.getenv('DATABASE_URL'):
     os.environ['DATABASE_URL'] = os.popen("heroku config:get DATABASE_URL -a found-your-stuff-api").read().strip() #seems jank
@@ -16,42 +13,6 @@ DATABASE_URL = os.getenv('DATABASE_URL')
 db = create_engine(DATABASE_URL, echo=True)
 Session = sessionmaker(bind=db)
 session = Session()
-Base = declarative_base()
-app = connexion.App(__name__, specification_dir='./')
-
-
-class User(Base):
-    __tablename__='users'
-    id = Column('id', Integer, primary_key=True)
-    email = Column('email', String, nullable=False)
-    password = Column('password', String, nullable=False)
-    name = Column('name', String)
-    phone_number = Column('phone_number', BigInteger)
-    contact = Column('contact', Boolean, nullable=False)
-    active = Column('active', Boolean, nullable=False)
-
-class Tag(Base):
-    __tablename__="tags"
-    id = Column('id', Integer, primary_key=True)
-    user_id = Column('user_id',Integer, ForeignKey("users.id"), nullable=False)
-    external_id = Column('external_id', UUID(as_uuid=True), nullable=False, default=uuid.uuid4, unique=True)
-    name = Column('name', String)
-    picture = Column('picture', String)
-    active = Column('active', Boolean, nullable=False)
-
-class Message(Base):
-    __tablename__="messages"
-    id = Column('id', BigInteger, primary_key=True)
-    tag_id = Column('tag_id',Integer, ForeignKey("tags.id"), nullable=False)
-    time_stamp = Column('time_stamp', DateTime, nullable=False)
-    body = Column('body', String, nullable=False)
-    picture = Column('picture', String)
-    read = Column('read', Boolean, nullable=False)
-
-def createTables():
-    Base.metadata.create_all(db)
-
-
 
 def getUserByExternalID(external_id):
     currentTag = session.query(Tag).filter(Tag.external_id==external_id).first()
@@ -108,10 +69,3 @@ def getAllUsersTags(user_guid):
     for tag in tags:
          listOfTags.append({"external_id":tag.external_id,"name":tag.name,"picture":tag.picture})
     return listOfTags
-
-app.add_api('openapi.yml')
-ENV = os.getenv('FYS_WORKING_ENV')
-if ENV:
-    if ENV.lower() == 'dev':
-        os.environ['FLASK_ENV'] = 'development'
-        app.run(port=8080, debug=True)
